Some customization points in BitType and preservation across generation.

+++ templates/BitType.php.tpl
@@ -200,6 +200,8 @@ class {/literal}{$render.class_name}{literal} extends {/literal}{$render.base_cl
 			$this->mErrors['title'] = 'You must enter a title for this {/literal}{$render.class|lower}{literal}.';
 		}
 
+		$this->customVerify($pParamHash);
+
 		return( count( $this->mErrors )== 0 );
 	}
 
@@ -212,8 +214,9 @@ class {/literal}{$render.class_name}{literal} extends {/literal}{$render.base_cl
 	function expunge() {
 		global $gBitSystem;
 		$ret = FALSE;
-		if( $this->isValid() ) {
+		if( $this->isValid() && $this->canExpunge() ) {
 			$this->mDb->StartTrans();
+			$this->customExpunge();
 			$query = "DELETE FROM `".BIT_DB_PREFIX."{/literal}{$render.class|lower}{literal}_data` WHERE `content_id` = ?";
 			$result = $this->mDb->query( $query, array( $this->mContentId ) );
 			if( LibertyMime::expunge() ) {
@@ -253,6 +256,8 @@ class {/literal}{$render.class_name}{literal} extends {/literal}{$render.base_cl
 		$bindVars = array();
 		array_push( $bindVars, $this->mContentTypeGuid );
 		$this->getServicesSql( 'content_list_sql_function', $selectSql, $joinSql, $whereSql, $bindVars );
+
+		$this->customGetList($pParamHash, $bindVars, $selectSql, $joinSql, $whereSql);
 
 		// this will set $find, $sort_mode, $max_records and $offset
 		extract( $pParamHash );
@@ -361,5 +366,43 @@ class {/literal}{$render.class_name}{literal} extends {/literal}{$render.base_cl
 {literal}
 		}
 	}
+
+	/* 
+	   -==-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
+	   User Modifiable Functions to add Custom Code
+
+	   Anything between the CUSTOM BEGIN and CUSTOM END
+	   comments will be preserved on regeneration of this
+	   file.
+	   -==-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
+	*/
+{/literal}
+	/* =-=- CUSTOM BEGIN -=-= */
+{if empty($customBlock)}
+	function customVerify(&$pParamHash) 
+	{ldelim}
+
+	{rdelim}
+
+	function canExpunge() 
+	{ldelim}
+	    return TRUE;
+	{rdelim}
+
+	function customExpunge() 
+	{ldelim}
+
+	{rdelim}
+
+	function customGetList(&$pParamHash, &$bindVars, 
+		 &$selectSql, &$joinSql, &$whereSql) 
+	{ldelim}
+
+	{rdelim}
+{else}
+{$customBlock}
+{/if}
+	/* =-=- CUSTOM END -=-= */
+{literal}
 }
 {/literal}
